+++ app/lib/l10n/app_localizations_en.dart
@@ -132,6 +132,12 @@ class AppLocalizationsEn extends AppLocalizations {
   String get go => 'Go';
 
   @override
+  String get previousPage => 'Previous page';
+
+  @override
+  String get nextPage => 'Next page';
+
+  @override
   String get pageNotFound => 'Page not found';
 
   @override